Clarify the relationship between the old and new algorithm
We've already explicitly described the policy graph as a more efficient
representation of the policy tree:

> The policy tree from {{RFC5280}} is the tree of all paths from the
> root to a leaf in the policy graph, so no information is lost in the
> graph representation.

Nonetheless, feedback on the list suggested this wasn't clear. Spell
this out more explicitly in both the text and the 5280 update, and give
explicit guidance for implementations that have to deal with
compatibility constraints.